build: Fix buildx use
Use `docker buildx build` explicitly rather than redefining
`CONTAINER_ENGINE` as `docker buildx`, as that caused docker runs be
executed as `docker buildx run`, which does not work.

Add `Makefile.docker` to the autogenerated dockerignore files, and
make `make` ignore it if missing.

Add support for detecting `ARCH` for docker builds. Defaults to empty
which is interpreted as build for the default platform of the current
builder (typically the host platform using the default
builder). `ARCH=multi` triggers multi-arch builds for all the
supported platforms, currently linux/amd64 and linux/arm64. Other
recognized values are `ARCH=amd64` and `ARCH=arm64` for single
platform builds.

When ARCH is specified and the current builder is the "default"
builder, a new cross-platform builder is created for all the supported
platforms.

Add support for `DOCKER_BUILDKIT_IMAGE` that allows dev to pass a
private image which is needed when using a self-signed docker registry
via `DOCKER_REGISTRY`.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>