cilium: encryption, insert "skip" rules to avoid ipsec policy
[ upstream commit 0b52fd76c0101e966d701c07ca174517948739e4 ]

In some configurations of tunnel mode, endpoint routes, etc. An encrypted
packet will be sent through the stack. If the encrypted packet does not
match a fwd policy rule AND it has previously been encrypted without passing
through a kernel boundary that clears the cb[] (on older kernels) or skb
extension (on newer kernels) fields the xfrm policy will miss and the
packet will be dropped.

This worked in the past because we mainly had older kernels where xfrm
stack used cb[] for metadata and we cleared the cb[] fields from BPF
stack. However, we do not clear skb extensions via BPF nor do we have
a BPF helper to do this yet. So in order to deal with this case we
add fwd xfrm policy rules so that when the routing stack detects the
skb extensions for xfrm and attempts to do a policy lookup a policy
is found and the packet is not dropped.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>